+++ src/main/java/utils/abstractFactory/AbstractFactory.java
@@ -1,4 +1,4 @@
-package utils;
+package utils.abstractFactory;
 
 import dao.IBaseDAO;
 import enums.DAOType;
